requirements.txt
flask
livestreamer
omxplayer-wrapper
youtube-dl

# sudo apt-get install python-dbus